Use int as index type everywhere except index buffer upload

+++ core/src/style/polygonStyle.cpp
@@ -40,7 +40,7 @@ void PolygonStyle::buildPoint(Point& _point, std::string& _layer, Properties& _p
 
 void PolygonStyle::buildLine(Line& _line, std::string& _layer, Properties& _props, VboMesh& _mesh) const {
     std::vector<PosNormColVertex> vertices;
-    std::vector<GLushort> indices;
+    std::vector<int> indices;
     std::vector<glm::vec3> points;
     std::vector<glm::vec2> texcoords;
     
@@ -69,7 +69,7 @@ void PolygonStyle::buildLine(Line& _line, std::string& _layer, Properties& _prop
 void PolygonStyle::buildPolygon(Polygon& _polygon, std::string& _layer, Properties& _props, VboMesh& _mesh) const {
     
     std::vector<PosNormColVertex> vertices;
-    std::vector<GLushort> indices;
+    std::vector<int> indices;
     std::vector<glm::vec3> points;
     std::vector<glm::vec3> normals;
     std::vector<glm::vec2> texcoords;
